+++ lib/idoklad/entities/currency.rb
@@ -4,6 +4,9 @@ module Idoklad
 
       self.entity_name = "Currencies"
 
+      def to_s
+        @table[:symbol]
+      end
     end
   end
 end