+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Currency entity model
 - ApiLimit model 
+- IssuedInvoice statuses
 ## [1.0.0] - 2019-06-02
 ### Added
 - subclasses structure of iDoklad entities